requirements.txt
django==5.0
djangorestframework==3.14.0
drf-yasg==1.21.7

psycopg2==2.9.9

python-dotenv==1.0.0

cryptography
paramiko

paho-mqtt